Move 5 functions from salt.utils
These functions are:

- salt.utils.ip_bracket -> salt.utils.zeromq.ip_bracket
- salt.utils.gen_mac -> salt.utils.network.gen_mac
- salt.utils.mac_str_to_bytes -> salt.utils.network.mac_str_to_bytes
- salt.utils.refresh_dns -> salt.utils.network.refresh_dns
- salt.utils.dns_check -> salt.utils.network.dns_check
