+++ salt/modules/event.py
@@ -14,6 +14,7 @@ import traceback
 # Import salt libs
 import salt.crypt
 import salt.utils.event
+import salt.utils.zeromq
 import salt.payload
 import salt.transport
 from salt.ext import six
@@ -60,7 +61,7 @@ def fire_master(data, tag, preload=None):
         # slower because it has to independently authenticate)
         if 'master_uri' not in __opts__:
             __opts__['master_uri'] = 'tcp://{ip}:{port}'.format(
-                    ip=salt.utils.ip_bracket(__opts__['interface']),
+                    ip=salt.utils.zeromq.ip_bracket(__opts__['interface']),
                     port=__opts__.get('ret_port', '4506')  # TODO, no fallback
                     )
         masters = list()
